refactor: Goods 컴포넌트명을 MainGoods로 변경
