+++ src/features/main/components/MainGoods/GoodItem/index.tsx
