+++ src/features/main/components/MainGoods/GoodItem/style.css.ts
