+++ src/features/main/components/MainGoods/style.css.ts
